+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# [3.8.3]
+* Fix for errors on a non scrollable editor.
+
+# [3.8.2]
+* Fix certain keys not working on web when editor is a child of a scroll view.
+
+# [3.8.1]
+* Refactor _QuillEditorState to QuillEditorState.
+
 # [3.8.0]
 * Support pasting with format.